Выносит тип в отдельный модуль
