+++ project/src/mocks/offers.ts
@@ -1,17 +1,4 @@
-type Offer = {
-  id: number;
-  mark?: string;
-  img: {
-    src: string;
-    width: number;
-    height: number;
-    alt: string;
-  };
-  price: number;
-  rating: number;
-  name: string;
-  type: string;
-}
+import {Offer} from '../types/types';
 
 const offers: Offer[] = [
   {
