+++ project/src/types/types.ts
@@ -1,0 +1,16 @@
+type Offer = {
+  id: number;
+  mark?: string;
+  img: {
+    src: string;
+    width: number;
+    height: number;
+    alt: string;
+  };
+  price: number;
+  rating: number;
+  name: string;
+  type: string;
+}
+
+export type {Offer}
